first cut at compiling with t_float set to double.  Using some ideas from Katja's fork of 0.43, but re-doing it with more willingness to let objects such as bonk~ simply use 'float' when suitable.
It might be desirable to revisit phasor~, etc., whose phase accuracy
limitations are noticeable in double (not so much in single precision).
In particular, the long-varispeed example now sounds OK with
just a phasor~ driving teh read, but the phase increment is
too quantized.

Scattered things appear to need a closer look (e.g., smeck
doesn't produce any outout).

+++ extra/expr~/vexp.c
@@ -1111,7 +1111,8 @@ abort();
         case ET_LB:
         default:
                 post_error((fts_object_t *) expr,
-                        "expr: ex_eval: unexpected type %d\n", eptr->ex_type);
+                        "expr: ex_eval: unexpected type %d\n",
+                            (int)eptr->ex_type);
                 return (exNULL);
         }
         if (!eptr[1].ex_type) {
@@ -1193,7 +1194,8 @@ abort();
         case OP_COMMA:
         case OP_SEMI:
         default:
-                post_error((fts_object_t *) expr, "expr: ex_print: bad op 0x%x\n", eptr->ex_op);
+                post_error((fts_object_t *) expr,
+                    "expr: ex_print: bad op 0x%x\n", (unsigned)eptr->ex_op);
                 return (exNULL);
         }
